correcciones constantes, MyButton y logHistory
