+++ screens/HomeScreen/constants.js
@@ -1,0 +1,3 @@
+
+export const NUM_ARRAY = ['7', '8', '9', '4', '5', '6', '1', '2', '3', '0'];
+export const OP_ARRAY = ['+', '-', 'x', '/'];
